requirements.txt
asgiref==3.7.2
Django==4.2.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
flake8==6.0.0
mccabe==0.7.0
Pillow==9.5.0
pycodestyle==2.10.0
pyflakes==3.0.1
sqlparse==0.4.4
typing_extensions==4.6.3
